usrloc(k): update for get_stat_val() type (cherry picked from commit dcc540027d3df990ba146dab6d5f86393b7da99a)

+++ modules_k/usrloc/dlist.c
@@ -558,7 +558,7 @@ void print_all_udomains(FILE* _f)
  */
 unsigned long get_number_of_users(void)
 {
-	int numberOfUsers = 0;
+	long numberOfUsers = 0;
 
 	dlist_t* current_dlist;
